+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to
 #### Changed
 - Use auto-resolution of library paths for tools
   - [#2181](https://github.com/iovisor/bpftrace/pull/2181)
+- Improve handling empty attach points
+  - [#2179](https://github.com/iovisor/bpftrace/pull/2179)
 
 #### Deprecated
 #### Removed
@@ -34,6 +36,8 @@ and this project adheres to
   - [#2138](https://github.com/iovisor/bpftrace/pull/2138)
 - Fix tools to work on new kernel versions
   - [#2136](https://github.com/iovisor/bpftrace/pull/2136)
+- Fix uprobe target resolution
+  - [#2180](https://github.com/iovisor/bpftrace/pull/2180)
 
 #### Tools
 #### Documentation